feat(platform): implement cursor disabling functionality in wayland platform layer

+++ src/events/platform_events.h
@@ -11,14 +11,14 @@ typedef struct {
 
 typedef struct {
 	EventCommon header;
-	int32_t x, y;
-	int32_t dx, dy;
+	double x, y;
+	double dx, dy;
 } MouseMotionEvent;
 
 typedef struct {
 	EventCommon header;
 	uint32_t button;
-	int32_t x, y;
+	double x, y;
 	uint32_t mods;
 } MouseButtonEvent;
